Move DataTypeSizeInfo to score/memory/shared
Moves the class to score/memory/shared and degrades it to an implementation detail.

score/memory/shared is currently the only user.
This class is dangerous since it gives no guarantee that the
size or alignment are actually correct for the associated type.
Users should prefer a RAII pointer style abstraction instead
(see scoped_function implementation details).

The safest approach is to remove it from the public API and
then clean up the current usages.
