+++ score/memory/BUILD
@@ -99,16 +99,6 @@ cc_library(
 )
 
 cc_library(
-    name = "data_type_size_info",
-    srcs = ["data_type_size_info.cpp"],
-    hdrs = ["data_type_size_info.h"],
-    features = COMPILER_WARNING_FEATURES,
-    tags = ["FFI"],
-    visibility = ["//visibility:public"],
-    deps = ["@score_baselibs//score/language/futurecpp"],
-)
-
-cc_library(
     name = "pmr_ring_buffer",
     hdrs = [
         "pmr_ring_buffer.h",
@@ -159,27 +149,11 @@ cc_test(
     ],
 )
 
-cc_unit_test(
-    name = "data_type_size_info_test",
-    srcs = ["data_type_size_info_test.cpp"],
-    features = COMPILER_WARNING_FEATURES + [
-        # These tests catch exceptions instead of using gtest EXPECT_DEATH so we disable aborts_upon_exception.
-        "-aborts_upon_exception",
-    ],
-    deps = [
-        ":data_type_size_info",
-        "@googletest//:gtest",
-        "@googletest//:gtest_main",
-        "@score_baselibs//score/language/futurecpp:futurecpp_test_support",
-    ],
-)
-
 cc_unit_test_suites_for_host_and_qnx(
     name = "unit_test_suite",
     cc_unit_tests = [
         ":endianness_unit_test",
         ":pmr_ring_buffer_unit_test",
-        ":data_type_size_info_test",
     ],
     test_suites_from_sub_packages = [
         "@score_baselibs//score/memory/shared:unit_test_suite",
