+++ score/memory/shared/data_type_size_info.cpp
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
-#include "score/memory/data_type_size_info.h"
+#include "score/memory/shared/data_type_size_info.h"
 
 namespace score::memory
 {
