+++ score/memory/shared/data_type_size_info.h
