+++ score/memory/shared/pointer_arithmetic_util.h
@@ -14,7 +14,7 @@
 // coverity[autosar_cpp14_m16_0_2_violation] false-positive: global namespace
 #define SCORE_LIB_MEMORY_SHARED_POINTER_ARITHMETIC_UTIL_H
 
-#include "score/memory/data_type_size_info.h"
+#include "score/memory/shared/data_type_size_info.h"
 
 #include "score/mw/log/log_types.h"
 #include "score/mw/log/logging.h"
